working edit test and changed the testng.xml file
