+++ testng.xml
@@ -4,6 +4,7 @@
   <test thread-count="5" name="Test">
     <classes>
       <class name="test.NewTestF"/>
+      <class name="test.NewTestFEdit"/>
       <class name="test.NewTestrEdit"/>
       <class name="test.NewTest"/>
       <class name="test.NewTestR"/>
